requirements.txt
Flask==3.0.2
pandas==2.2.1
scikit-learn==1.4.1.post1
joblib==1.3.2
python-dotenv==1.0.0
gunicorn==21.2.0
